refactor(GameCreator): move functionality to gameLogic dir
